requirements.txt
fastapi
uvicorn
jinja2
numpy
pyyaml
pyserial